# src/docs/openapi-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/media/secure-url/{media_id}": {
      "get": {
        "operationId": "MediaController_getSecureFileUrl",
        "summary": "",
        "description": "Retrieves a secure URL for a specific media file.",
        "parameters": [
          {
            "name": "media_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Secure URL retrieved successfully."
          }
        },
        "tags": [
          "media"
        ]
      }
    },
    "/media/url/{media_id}": {
      "get": {
        "operationId": "MediaController_get_Publice_File_Url",
        "summary": "",
        "description": "Retrieves a public URL for a specific media file.",
        "parameters": [
          {
            "name": "media_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Public URL retrieved successfully."
          }
        },
        "tags": [
          "media"
        ]
      }
    },
    "/media/delete/{media_id}": {
      "delete": {
        "operationId": "MediaController_delete_media_file",
        "summary": "",
        "description": "Deletes a specific media file by ID.",
        "parameters": [
          {
            "name": "media_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Media file deleted successfully."
          }
        },
        "tags": [
          "media"
        ]
      }
    },
    "/categories": {
      "get": {
        "operationId": "CategoryController_getAll",
        "summary": "",
        "description": "Retrieves a list of all categories.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of categories retrieved successfully."
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "tags": [
          "categories"
        ]
      },
      "post": {
        "operationId": "CategoryController_create",
        "summary": "",
        "description": "Creates a new category.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Category created successfully."
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "tags": [
          "categories"
        ]
      }
    },
    "/categories/{id}": {
      "get": {
        "operationId": "CategoryController_findOne",
        "summary": "",
        "description": "Retrieves a specific category by ID.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Category retrieved successfully."
          },
          "404": {
            "description": "Category not found."
          }
        },
        "tags": [
          "categories"
        ]
      },
      "put": {
        "operationId": "CategoryController_update",
        "summary": "",
        "description": "Updates a specific category by ID.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Category updated successfully."
          },
          "404": {
            "description": "Category not found."
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "tags": [
          "categories"
        ]
      },
      "delete": {
        "operationId": "CategoryController_remove",
        "summary": "",
        "description": "Deletes a specific category by ID.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Category deleted successfully."
          },
          "400": {
            "description": "Bad request."
          },
          "404": {
            "description": "Category not found."
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "tags": [
          "categories"
        ]
      }
    },
...
    "/webhook/status-update": {
      "post": {
        "operationId": "WebhookController_handleStatusUpdate",
        "summary": "",
        "description": "Handles status update events from external services.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Notification payload for status updates.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/notifyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Webhook processed successfully."
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "tags": [
          "webhook"
        ]
      }
    },
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "summary": "",
        "description": "Registra un nuevo usuario en el sistema",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuario registrado exitosamente"
          },
          "400": {
            "description": "Datos de registro inválidos"
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "summary": "",
        "description": "Inicia sesión para un usuario existente",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Inicio de sesión exitoso"
          },
          "401": {
            "description": "Credenciales inválidas"
          }
        },
        "tags": [
          "auth"
        ]
      }
    }
  },
  "info": {
    "title": "Farmers Marketplace API",
    "description": "API documentation for the Farmers Marketplace platform",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/api/v1"
    },
    {
      "url": "/api/v1"
    },
    {
      "url": "/api/v2"
    },
    {
      "url": "/api/v1"
    },
    {
      "url": "/api/v2"
    }
  ],
  "components": {
    "securitySchemes": {
      "access-token": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization"
      }
    },
    "schemas": {
...
      "CreateCategoryDTO": {
        "type": "object",
        "properties": {}
      },
      "UpdateCategoryDTO": {
        "type": "object",
        "properties": {}
      },
...
      "notifyDto": {
        "type": "object",
        "properties": {}
      },
      "RegisterDto": {
        "type": "object",
        "properties": {}
      },
      "LoginDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}